Group logic together in extensions
